split sqllogic
