+++ testcase/sqllogictest/client.go
